refactor(cli): extract git-client boundary module from github-skills
Split github-skills.ts into domain logic and git CLI operations:
- Create lib/external/git-client.ts for downloadGitHubSkill and
  downloadGitHubDirectory (all child_process.exec git operations)
- Keep github-skills.ts as pure domain logic (URL parsing, SKILL.md
  validation/parsing) with re-exports for backward compatibility
- Update compose test to mock git-client (full mock) instead of
  github-skills (partial mock with importOriginal)

Closes #3007
